ml-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.0.1
opencv-python==4.8.1.78
ultralytics==8.0.196
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
pydantic==2.4.2
python-dotenv==1.0.0
aiofiles==23.2.1
